Fix ensure_versioned_database_exists to not open a file for writing
